requirements.txt
Django==4.2.7
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
pillow==10.1.0
django-ckeditor==6.7.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-extensions==3.2.3
python-decouple==3.8
django-bootstrap5==23.3 